Create database/seeders/book.seeder.ts
