--- src/database/models/book/interfaces/IBookDocument.ts
@@ -1,13 +1,0 @@
-import { Document } from 'mongoose';
-import { Book } from '../../../../common/constants/book';
-
-export interface IBookDocument extends Document {
-    readonly id: string;
-    readonly ownedBy: string;
-    readonly borrowedBy: string | null;
-    readonly title: string;
-    readonly description: string;
-    readonly genre: Book.GENRE;
-    readonly picture: string;
-    readonly addedAt: number;
-}
