--- src/database/models/book/interfaces/IBookRepository.ts
@@ -1,5 +1,0 @@
-import { Book } from '../book';
-
-export interface IBookRepository {
-    create(data: Partial<Book>): Promise<Book>;
-}
