--- src/database/models/book/seeder/seeder.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BookSeeder } from '.';
-import { BookRepositoryModule } from '../book.repository.module';
-
-@Module({
-    imports: [BookRepositoryModule],
-    providers: [BookSeeder],
-    exports: [BookSeeder],
-})
-export class BookSeederModule {}
